requirements.txt
ebooklib
google-api-python-client
bs4
